Move Workspace class out of pouch module
